ground truth nsrts for cover
