+++ predicators/ground_truth_models/__init__.py
@@ -1,12 +1,14 @@
 """Implements ground-truth NSRTs and options."""
 
 import abc
-from typing import Set
+from typing import Dict, Set
 
 from predicators import utils
 from predicators.envs import BaseEnv, get_or_create_env
+from predicators.ground_truth_nsrts import \
+    get_gt_nsrts as deprecated_get_gt_nsrts
 from predicators.settings import CFG
-from predicators.structs import ParameterizedOption
+from predicators.structs import NSRT, ParameterizedOption, Predicate, Type
 
 
 class GroundTruthOptionFactory(abc.ABC):
@@ -22,6 +24,24 @@ class GroundTruthOptionFactory(abc.ABC):
     @abc.abstractmethod
     def get_options(env_name: str) -> Set[ParameterizedOption]:
         """Create options for the given env name."""
+        raise NotImplementedError("Override me!")
+
+
+class GroundTruthNSRTFactory(abc.ABC):
+    """Parent class for ground-truth NSRT definitions."""
+
+    @classmethod
+    @abc.abstractmethod
+    def get_env_names(cls) -> Set[str]:
+        """Get the env names that this factory builds NSRTs for."""
+        raise NotImplementedError("Override me!")
+
+    @staticmethod
+    @abc.abstractmethod
+    def get_nsrts(env_name: str, types: Dict[str, Type],
+                  predicates: Dict[str, Predicate],
+                  options: Dict[str, ParameterizedOption]) -> Set[NSRT]:
+        """Create NSRTs for the given env name."""
         raise NotImplementedError("Override me!")
 
 
@@ -43,6 +63,42 @@ def get_gt_options(env_name: str) -> Set[ParameterizedOption]:
     for option in options:
         option.params_space.seed(CFG.seed)
     return options
+
+
+def get_gt_nsrts(env_name: str, predicates_to_keep: Set[Predicate],
+                 options_to_keep: Set[ParameterizedOption]) -> Set[NSRT]:
+    """Create ground truth options for an env."""
+    # This is a work in progress. Gradually moving NSRTs into env-specific
+    # files; ground_truth_nsrts.py will be deleted.
+    env = get_or_create_env(env_name)
+    env_options = get_gt_options(env_name)
+    assert predicates_to_keep.issubset(env.predicates)
+    assert options_to_keep.issubset(env_options)
+    for cls in utils.get_all_subclasses(GroundTruthNSRTFactory):
+        if not cls.__abstractmethods__ and env_name in cls.get_env_names():
+            factory = cls()
+            # Give all predicates and options, then filter based on kept ones
+            # at the end of this function. This is easier than filtering within
+            # the factory itself.
+            types = {t.name: t for t in env.types}
+            predicates = {p.name: p for p in env.predicates}
+            options = {o.name: o for o in env_options}
+            nsrts = factory.get_nsrts(env_name, types, predicates, options)
+            break
+    else:
+        # In the final version of this function, we will instead raise an
+        # error in this case.
+        nsrts = deprecated_get_gt_nsrts(env_name, predicates_to_keep,
+                                        options_to_keep)
+    # Filter out excluded predicates from NSRTs, and filter out NSRTs whose
+    # options are excluded.
+    final_nsrts = set()
+    for nsrt in nsrts:
+        if nsrt.option not in options_to_keep:
+            continue
+        nsrt = nsrt.filter_predicates(predicates_to_keep)
+        final_nsrts.add(nsrt)
+    return final_nsrts
 
 
 def parse_config_included_options(env: BaseEnv) -> Set[ParameterizedOption]:
