+++ predicators/ground_truth_models/cover/__init__.py
@@ -1,5 +1,6 @@
 """Ground-truth models for cover environment and variants."""
 
+from .nsrts import CoverGroundTruthNSRTFactory
 from .options import CoverGroundTruthOptionFactory
 
-__all__ = ["CoverGroundTruthOptionFactory"]
+__all__ = ["CoverGroundTruthOptionFactory", "CoverGroundTruthNSRTFactory"]
