+++ tests/approaches/test_pg4_approach.py
@@ -6,7 +6,7 @@ from predicators import utils
 from predicators.approaches import ApproachFailure, ApproachTimeout
 from predicators.approaches.pg4_approach import PG4Approach
 from predicators.envs import create_new_env
-from predicators.ground_truth_nsrts import get_gt_nsrts
+from predicators.ground_truth_models import get_gt_nsrts
 from predicators.structs import LDLRule, LiftedAtom, LiftedDecisionList, Task
